CreateTiledChannelGraph: Show the headstage as well
We now add an additional free axis which will be used for labeling the
DA and AD channels with their headstage. And also move
the current axis labels closer to the axis.

If both are shown we use the axis range of both (including the distance
between them) otherwise we only use the respective axis lengths.

This also requires us to remove any remaining free axis before we start,
this is done by RemoveFreeAxisFromGraph.