Extended PopupMenus procedures to replace long popup menus
The procedures keep a compatible upgrade, where long popupmenus
get replaced by a button and previous proc definition move to the
buttons userdata.
After the menu is clicked the old proc gets called as if it would be a
popupmenucontrol.

The inline documentation includes three examples.

Additional features allow to define the menu item distribution in
the sub menus automatically or manually. For automatic distribution
and sub menu naming utility functions are provided with various
methods.