In DB SettingsHistory replaced popupmenu LNB selection with extended popupmenu
Two popup menus with very long menu item lists in the data browser settings
history window are replaced by a single menu using the new popup extension.

The setup for changed menu uses waves only now. The utility function
GetLabNotebookSortedKeys was used nearly exclusively for the DB and was
updated to use waves instead of lists. Another function for unsorted
keys was added as well: GetLabNotebookKeys

A new utility function in the DB was added: DB_GetLNBKeyEntryType
It retrieves the type of a LNB entry and returns one of the
LNBEntryTypes constants.