Labnotebooks: Fix too short dimension labels
Since forever we've limited the length of dimension labels set via
SetDimensionLabels to 31 bytes.

And we already have one labnotebook entry "Slow current injection settling time" which is cut off. This breaks querying the entry from the labnotebook as we always use the dimension label.

In a previous commit we already raised the
MAX_OBJECT_NAME_LENGTH_IN_BYTES, so now we just have to fixup old
labnotebook waves.