ConvertListToRegexpWithAlternations: Fix returned regexp
We need to remove the trailing "|" so that the returned regexp does not
match the empty string.

Broken since d13a6b72 (GetListOfObjects: Rewrite and speed it up a lot,
2020-03-26).

And while we are at, we also encapsulate the whole in non-matching
brackets to avoid any problems when the result value is used in another
regular expression.