CreateTiledChannelGraph: Update tgs.displayXXX if required
We check that the display option for each channel is enabled and that we
have at least one channel of the given type.

But if that is not the case we should set the display option to zero as
we want to query that later on without checking the number of channels
again.