DAP_CheckStimset: Fix checking required analysis parameters
We have a semicolon separated list in reqNames and also the indizes in
the loop header were wrong.

Bug introduced in ff7dd72a (Analysis parameters: Add method to request
the types as well, 2018-05-13).

The bug manifested itself through only checking the 6th required
parameter and beyond.

Embarissingly the tests expected the broken behaviour. This is now fixed
as well.