Add a generic way to check analysis parameters
Until now every analysis function had to check the passed analysis
parameters on each call. This also results in problems if you e.g.
realize that you have invalid analysis parameters in the mid sweep
event.

We now support a user supplied `_CheckParam` function which can check if
the passed analysis parameter is valid or not. Only required and
optional parameters from `_GetParams` are passed to the check function.

Invalid analysis parameters can not be stored in stimsets and stimsets
with these can also not be used for DAQ. The check function must also
not abort on a given parameter, as that is also treated as an error.