DAP_CheckStimset: Fix analysis parameter content checking for waves
We must now pass in the list of required parmeters with types into
AFH_GetAnalysisParamWave/AFH_GetAnalysisParamTextWave but the serialized
parameters from the stimset.

Bug introduced in ff7dd72a (Analysis parameters: Add method to request
the types as well, 2018-05-13).

To make the code clearer reqParams was also renamed to
reqNamesAndTypesFromFunc.