NWB_AppendSweepLowLevel: Handle broken data better
We have one file, H18.26.404.11.05.01.01.pxp, which has in sweep 9 two
active headstages but HS2 does miss the corresponding AD channel.

This is not a valid setup and currently bugs out due to
CreateLBNUnassocKey getting a NaN channel number.

We can workaround that easily by treating the headstage as inactive and
exporting the channel as unassociated channel.