Make cursor saving and restoring faster
We don't need to save the cursor info text wave if there are no cursors
on the graph. If there are cursors on the graph this can be checked with
GraphHasCursors() which is faster than just creating it.

This approach also allows us to not query the list of traces in
RestoreCursors() when there is no cursor info.