OverlaySweeps: Don't replot when we want to highlight a sweep
Since forever we handled sweep highlightning (which is used during
overlay sweeps editing of the to-be-removed headstages) inside
CreateTiledChannelGraph which required a full replot.

This is quite wasteful.

The new implemntation of OVS_HighlightSweep does now handle the
translucency application, does not require the dimension label of the
listbox wave for intermediate storage, and is of course much faster.