GetPlainSweepList: Return a numeric wave with just the sweep numbers
No callers are interested in the current return value of a string with
"Sweep_XXX" formatting. So by just returning a numeric wave with the
sweep numbers the call sites will have less code.

This also fixes a theoretical bug in DB_FirstAndLastSweepAcquired as it
assumed that the sweep numbers are continous. We are now returning the
first and last values instead.