Databrowser/SweepBrowser: Add generic function for LBN fetching
We used to always branch out when fetching the labnotebook waves. This
is error prone, duplicated code and can be done better.

We now have BSP_GetNumericalValues and BSP_GetTextualValues which return
a wave with the labnotebooks of all requested sweeps. For easier usage
it returns the labnotebook wave directly when only the labnotebook for a
single sweep is requested.

The sweep browser labnotebook functions now also behave the same way.

This also allows us to remove DB_UpdateOverlaySweepWaves.