Add scale bars to the PA plot
This adds X and Y scale bars to the non-diagonal sets of the PA plot.
The Y scale has the given length from the GUI, the x scale is a multiple
of 5ms long (and closest to 10% of the full scale).

We update the scale bar positions when the axis ranges change. The free
axis called "fakeAxis" and its hook function PA_AxisHook is used for
that.