PA plot: Overlay the y-axis and the y scale bars exactly
We now start the y axis 1% (0.01) before the x-axis (in horizontal
coordinates). Thus the x-axis needs to be shifted to the right as well
for that.

The y scale bar now uses the same position as the y-axis, but that
required to use prel coordinates for the y-scale bar x coordinates.