MIES_fWaveAverage: Rework it
We used to do only small changes in order to facilitate upstreaming the
results. But that makes chaning the code much harder. Therefore we now
remove things we don't need, error waves, x waves and are also free to
change the API.

The function now takes a wave reference wave and returns a free wave. In
addition the caller can now choose to ignore/respect NaNs in the input
data and the resulting data type can be given as well.

Finally we add tests so that this stays working.

The function CalculateAverage also requires rework to cope with the
returned free wave instead of the in place creation of the new average.

The call sites of PA_Average also got faster as the calls to
WaveRefWaveToList are not necessary anymore.