CalculateAverage: Update average wave when fetching from the cache
The current code is unnecessarily complex. We can just take the wave
from the cache, skipping the internal duplicate, and duplicate it into
the correct place.

This has the advantage that this does not require a KillOrMoveToTrash
and keeps the wave references intact, which might be useful when the
average wave is shown in a graph.