Avoid regenerating the Labnotebook cache from scratch if possible
In the initial introduction of the labnotebook cache waves in 7aee1b63
(Add a cache for the labnotebook queries, 2018-05-28) we already noted
that we completely flush the cache when the labnotebook changes.

This is correct, but slower than it could be.

The following assumes that no sweep rollback happened. When sweep
rollback happens we do the update as usual, for that we track the sweep
rollback count.

A general property of the labnotebook is that we only ever append
entries. This means entries in the labnotebook caches would need an
update if the additions to the labnotebook invalidated some entries.

We have to distinguish the free cache waves:
- GetLBRowCache: This only holds valid entries, but we could still add
  an entry to the currently last sweep, so we need to invalidate the
  current sweep.
- GetLBIndexCache: This holds also missing values, so we need to convert
  all entries which now might not be missing anymore. As the row indizes
  are sweep numbers, we can just set everything after the currently
  acquired sweep to LABNOTEBOOK_MISSING_VALUE
- GetLBNidCache: We don't do incremental update here as we would need to
  touch multiple entries as the new data touches multiple (or even all)
  sweeps.

Once that update was done we also set the labnotebook wave modification
count in our cache wave note to the current entry, as by-definition our
cache is now up to date with the labnotebook values waves.

The initial sizes of the cache waves were also a bit lowered to allow
for faster incremental updates.