PA_ShowPulses: Calculate average in parallel if possible
When the pulses change we always need to recalculate the average. By
doing that in parallel this is faster than the current serial approach.

We store the so calculated average wave in the wave cache as well for
later access.

We also switch to no-ignoring NaNs and 32bit floating point values for
the average to make it faster as well.