Wavebuilder: Avoid ambigous stimset creation waves
Since forever we have to function for creatim stimulus sets.
WB_CreateAndGetStimSet is the public function which permanently creates
the stimset and returns a wave reference to it. And WB_GetStimSet which
returns a free wave for the current wavebuilder stimulus set or the
given stimset.

This is difficult to understand and also the author of this commit
struggles sometimes to quickly pick the right one.

Introduce WB_GetStimSetForWaveBuilder() and make WB_GetStimSet static,
therefore WB_CreateAndGetStimSet is now the only sensible choice.

Close #532.