DB_ResetAndStoreCurrentDBPanel: Allow to use the same panel multiple times
We clear the panel version before regenerating the window recreation
macro. This is good and intendent but currently has the drawback that
you can not use the panel twice for that.

Let's just add the current panel version if it is too low, to make it
more convenient to use.