NumericWaveToList: Workaround %d output bug
Since 378d514f (NumericWaveToList: Make it faster, 2020-08-14) we are
using wfprintf in NumericWaveToList.

Unfortunately this does not work for %d with data beyond 2^31 -1. Until
we have a fix from WaveMetrics we can rely on, let's use ".15g" instead
for all callers.

We also raise the average key version as erroneous cache entries was
what triggerd this bug hunt.