PA Plot: Reimplement time alignment
We are currently using the generic TimeAlignmentIfReq function for time
alignment. This has the disadvantage that it requires our pulses to be
plotted. We also have very special needs for the PA time alignment,
which are difficult to get fast in TimeAlignmentIfReq.

Therefore we've now reimplemented that in PA_AutomaticTimeAlignment
exploiting our properties and set waves which do not require plotted
data.