OVS_UpdatePanel: Make it run faster
OVS_UpdatePanel is always called when we have acquired new sweeps and a
databrowser is open.

This function currently iterates over all sweeps and queries various
labnotebook entries for populating the sweep selection choices for the
popupmenu in the overlay sweeps tab.

This is a rather expensive task which only needs to be done if the user
chooses the popup menu the first time after changes.

This is now implemented.