Databrowser/SweepBrowser: Always create backup waves when splitting the sweep waves
Our current approach is to create backup wave before doing write
operations on the waves. This is error prone.

By doing the backup when creating the 1D waves from the 2D sweep waves,
we ensure that we always have a backup.

For sweeps loaded from NWB files we don't have 2D sweep waves, so we
create the backup when loading them in AB_LoadSweepFromNWBgeneric.