PA_CreateAndFillPulseWaveIfReq: Fix failed pulse detection
Since cae93d42 (MIES_PulseAverage.ipf: Enhance and add failed pulse
detection, 2020-07-23) we support failed pulse detection.

For failed pulse detection we add wave note entries in
PA_MarkFailedPulses. But these wave note entries land in the backup wave
when the single pulse wave is recreated.

By clearing the wave note we avoid that. The bug manifests itself by
marking pulses as failed which actually passed.

But we must keep the wave version on clearing the wave note as otherwise
the single pulse wave is recreated on the next call of
PA_CreateAndFillPulseWaveIfReq.