Pulse averaging: Calculate the pulse length better
We always used the number of requested pulses for the difference
calculation, and if we can not do that, for zero or one pulse, we just
return the fallback length.

But taking the number of requested pulses here always returns the
fallback length if only one pulse is shown. So let's use the number of
available pulses instead.