Fix incremental LBN row cache
In 3d4e6450 (Avoid regenerating the Labnotebook cache from scratch if
possible, 2020-08-10) we added incremental labnotebook cache update.

This did not work in the edge case that between writing a new LBN entry
of an existing sweep and creating a new sweep no write occured. Then the
search for the last value of the previous sweep fails as the range
was only invalidated for the very last sweep.

We now invalidate ranges for the last and its predecessor.

We also fix a potential bug as EnsureLargeEnoughWave was not using the
correct initial value.