OVS_EndIncrementalUpdate: Handle no sweeps better
In case we have no sweeps we currently bug out as listBoxWaveAfter and
listSelWaveAfter will not have the correct dimension labels.

These are getting lost when we duplicate the waves from (0, 2) size as
the target will only have (0).

The solution is two fold. For once we do a full update if the now don't
have any sweeps displayed. This is cheap.

And we also copy the dimension labels from the after waves to the before
waves as the before waves don't have the correct dimension labels.