Labnotebook: Use the same timestamp for all layers
Since forever we called DateTime/DateTimeInUTC() for each layer. But
this can lead to  slightly different values as they are called multiple
times.

This bug does not result in issues when reading as we don't claim any
properties about the time stamps.

Call them once and then set the values.