AB_ScanFolder: Prevent loading the currently written to NWB file
This does not work as we get a HDF5 library error when we try to read
with one file descriptor from it and write into it with another one.

Even if we used the same file descriptors for it, we would need updating
support in the analysis browser to make it really worthwhile using it.

So now we ignore the currently exported into file and warn the user.