RA_SkipSweeps: Fix various corner cases
The current code did not expect that you call first without
limitToSetBorder=0 and then with limitToSetBorder=1.

It was also not working for -inf and did not adjust the active set count
correctly when limitToSetBorder=0.

These bugs are now fixed. The solution is to first calculate a skipCount
which is limited to the set borders if requested. And when a new count
is calculated using RA_SkipSweepCalc and the so calculated count offset
is used for adjusting count and the activeSetCount.