tools/unit-testing/autorun-test.sh: Add option to timeout the execution
On a certain class of errors IP does currently display a dialog. This
results in the hung build killer from bamboo killing the shell process
[1], but that does not include the IP instance which was started [2].

We therefore need to handle that ourselves. This solution here is
modelled after the initial solution from ab4bb524 (Switch to IP7 64bit
for unit and compilation testing, 2017-08-20).

[1]:
https://confluence.atlassian.com/bamkb/script-builder-doesn-t-kill-child-processes-189891893.html
[2]: https://stackoverflow.com/a/49378862