Make more analysis function routines available in a generic manner
We used to have the PSQ_CreateLBNKey and PSQ_MapFunctionToConstant functions also in
variants for MSQ. This approach does not work when we want to add
dashboard support also for MSQ function as we would constantly need to
check both of them.

Generalize both routines to CreateAnaFuncLBNKey and
MapAnaFuncToConstant.

We also had to change the MSQ_DA_SCALE and MSQ_FAST_RHEO_EST constants
so that these don't overlap with the PSQ ones.