Tests: Remove CheckLength tests
In f88b7cb8 (Add test to check that all generated keys from PSQ_CreateLBNKey are short enough, 2018-05-10)
we added a check so that the Labnotebook keys for the patchseq analysis
functions are not too long.

But since 9c91f169 (Raise the general object name limit to 255 bytes, 2020-04-09) we raised the limit to 255 bytes.

Let's remove that test as it is not likely that we use labnotebook keys
longer than 255 bytes.