NumericWaveToList: Avoid RTE on empty wave
Since 378d514f (NumericWaveToList: Make it faster, 2020-08-14) we are
using wfprintf in NumericWaveToList.

wfprintf does not touch the passed output string when the input wave is
empty.

This makes sense but does result in NumericWaveToList returning an null
string and thus generating a RTE.

Intializing the list string with an empty string fixes the problem.