Adapted MIES_fWaveAverage to support incremental averaging
Generally for incremental averaging the function has to return
the sum per row as well as the counts per row additionally.
For this the new optional parameter getComponents was introduced.

To keep auto indexing calls to MIES_fWaveAverage possible, instead of
using MRS the function now returns a wave reference wave with 3 rows.

The first row is the average wave as before.
The second row are the sums and the thrid the counts.
The second and third row are only setup with valid return data when
getComponents = 1 was used.

Having the sums per row and counts per row, it is also possible to call the
averaging incremental. Previous averaging data can be given by the optional
prevAvgData wave reference parameter.
The first entry must contain the sums and the second the counts per row.

This data is then included in the averaging of the yWaves.