PA_ZeroPulses: Don't recalculate Min/Max if no zeroing was done
In 6864489f (PA plot: Add data minimum/maximum to the wave note, 2020-08-03)
we added a multithread statement for ZeroWave and a call next to it for
PA_UpdateMinAndMax.

With using + both were always executed. But we only need to update the
Min/Max if we really did zero the wave.

Let's change it to that.