AD_UpdateAllDatabrowser: Don't bug out when we have PA plot windows open
We used to just call AD_Update for all windows which start with "DB_".
But since 32d03db7 (PA plot: Allow one PA plot per browser window,
2020-08-05) we prefix the PA plot windows with the databrowser name as
well as the PA plot windows are now per databrowser.

And that broke

Use the now available BSP_GetBrowserType to check if it is a
databrowser or sweepbrowser window.