Hardware Tests: Avoid bugging out with new UTF version
The new UTF version now bugs out if the wave type of the data generator
function is not the correct one compared to the test case.

And since d1507158 (HardwareTests use now multi data test case feature
to run for each device, 2019-04-26) we return empty floating point waves
when we don't want to run the tests for this setup.

But we need to return empty text waves.