WaveVersioning: Use explicit wave types
By default Make does use 32-bit real values (/R). This is not obvious
and can lead to bugs as that default value might not be the right choice.

Let's use an explicit wave type in the documentation as that is what we
are going to use for all wave getters.