Wavebuilder: Use WP and SegWvType with double precision for new waves
Since forever we are using single-precision waves for storing the
stimulus set parameters (WP) and segment parameters (SegWvType).

This limits the precision of the parameters to six digits. If we would
start over we would just use double precision for both waves
and that's it. But doing that now breaks backwards compatibility.

The biggest loss of precision is due to us using a copy of WP
for calculating the new values for the sweeps in WB_AddDelta.

So we use double precision for new waves but keep existing waves in
single precision.