Hardware Tests: Use floating point comparison with tolerance for gain
Since the introduction of double precision for the GetChanAmpAssign wave
we now have small deviations in the comparison of the stored gain in NWB
and the one in the labnotebook.

The reason is that we store in NWB only a float but in the
labnotebook we have double.

The gain values are 0.000500000023748726 (NWB, float) and 0.0005
(labnotebook, double) so we don't loose any relevant precision.
Therefore using CHECK_CLOSE_VAR is more appropriate.