OVS_ChangeSweepSelectionState: Avoid always selecting the first sweep
Since the enhancement in 0a71aec8 (OVS_ChangeSweepSelectionState:
Enhance it, 2020-10-23) we use indices as an index wave.

The help topic "Indexing with an index wave" was unfortunately misread
by the author of this commit. We don't need a two point 1D wave, but a
two point 2D wave, where we have all columns of listboxWave set.

Fix that.