PA_DrawScaleBarsHelper: Draw scale bar lengths if different
We have one x axis and scale bar for each region (column). In order to
not clutter the image we only add one x scale bar to the bottom right
set and one y scale bar to the bottom left set.

We allow users zoomming in and out into each axis, and then
dynamically resize the scale bars to a fitting length.

This makes it really hard to know the exact physical length when
zooming.

We solve that by remembering the initial bar length and always draw
the bar length if it is different from the inital value. At least for
the x scale bars.

The y case is more complicated as we have a user setting as well. So we
prioritize the user setting and only choose a nice length if the user
zommed in or out using the mouse wheel.

And we reset the user choosen scale bar lengths when we modify the plot
programmatically or the user auto scaled the axis.