PSQ_Rheobase: Use better floating point comparison routines and fix test
Since the move to more double precision waves in the last commits, the
test PSQ_RB2 is failing.

The reason is that it stops earlier now as it decides that
PSQ_RB_MAX_DASCALE_DIFF is reached earlier.

This is due to floating point rounding issues.

After the 5th sweep we have in master

initialDAScale 0.000000000070000000000000003843
DAScale        0.000000000129999999999999990519
diff           0.000000000059999999999999986676

And that this diff is smaller than 60e-12, PSQ_RB_MAX_DASCALE_DIFF, so
we acquire one sweep more.

But now we have

initialDAScale 0.000000000070000000000000003843
DAScale        0.000000000130000000000000016368
diff           0.000000000060000000000000012526

and that is larger so we stop.

We want to avoid issues like that in principle so let's use the
CheckIfClose routines which take care of that for us.

This requires to adapt the tests as well. This is no problem regarding
backward compatibility as we now stop earlier if we should, so this is a
real bon.