PA_ShowImage: Fix incremental update and failed pulses highlightning
When we do incremental update via POST_PLOT_ADDED_SWEEPS we need to find
the new sweep in setIndizes and then copy all pulses from that entry to
the last one.

But this means requiredEntries as the column index past the last valid entry does
not change.

And that also means that PA_HighligthFailedPulsesInImage must start
looking at the very first pulse and not at firstPulseIndex as we remove
all failed pulses rectangles on each new graph.

Close #690
Close #678