PA_ShowImage: Avoid empty lines when incrementally adding sweeps
When we add new data using the POST_PLOT_ADDED_SWEEPS fast path, we are
currently using a from scratch calculated singlePulseColumnOffset value.
But that value might differ from the last time we wrote into img.

Threfore we need to store the singlePulseColumnOffset value in the wave
note, and reuse that if it is present.