Databrowser: Fix overlaying sweeps doing an incremental update for existing sweeps
Recipe for reproduction:
- Open Databrowser
- Open PA plot
- Enable OVS

You know have duplicated traces in the PA plot. The reason is the
incremental OVS update introduced in ffca2076 (Add incremental update
support for overlay sweeps, 2020-07-15). The call to OVS_UpdatePanel in
BSP_CheckProc_OverlaySweeps does not recognize any new sweeps, the
selected sweeps list will be empty, but we do a full update
including calling UpdateSweepPlot. But as we have no sweeps selected we
take the early return in DB_UpdateSweepPlot, and thus don't update the
PA plot as well.

Then we select a sweep with OVS_ChangeSweepSelectionState, this new
sweep is found with the incremental update and we plot the new data from
that sweep in the PA plot.

But as that data is already shown in the PA plot we now have
duplicated traces.

The fix is to make the sweep selection when nothing is selected already
in OVS_UpdatePanel, thus replotting everything correctly, and skipping
the incorrect update later on.

Close #709.