WIP: RA_SkipSweeps: Document sweep skipping when requested
For post-mortem data analysis it is difficult to judge why a
certain sweep was not acquired.

With the new "Sweep Skipping" labnotebook entry, this is now documented
for interactive use with the buttons in the DAEphys panel.

For documenting that we need to read back the current value and add that
on top of the new skip count. The reason is that users might press the
skip sweep buttons multiple times during one sweep. And the way our
labnotebook works we only get the last value (this is intended). So we
need to read back the current value and add the new value on top of
that.

This is also one of the first users of the acquisition state variable
via AS_GetCurrentSweepNumber.

Close #718.