PA_DrawScaleBarsHelper: Avoid requiring that the axis are already scaled
We currently query AxisInfo for getting the x axis start value. But that
requires that the graphs was already drawn.

We can avoid that by setting the x axis start position in
PA_LayoutGraphs as user data and query that in PA_DrawScaleBarsHelper.
This also requires to call PA_LayoutGraphs earlier as
PA_DrawScaleBarsHelper.