ZeroWaveImpl: Reimplement it with less complicated code
When we are taking the numeric derivative and numeric integral of the
data with end up with the starting data +/- a constant.

And it happens that the implementation in IP chooses the constant such
that in most cases the first values is zero.

So we can just do that as it is much faster that way.

Close #705.