RemoveUnusedRows: Handle non-existing NOTE_INDEX gracefully
When exporting experiments to NWB prior to e96222a6 (TPStorage: Unify on
NOTE_INDEX, 2018-10-17) the TPStorage waves don't have a NOTE_INDEX key.
And the assertion triggers as index is NaN. This was introduced in
c5b1ad32 (NWB_AddDeviceSpecificData: Only write the used rows into the
NWB files, 2020-09-24).

Now we could add a fallback path to try "TPCycleCount" instead. But that
would add a workaround to an optional optimization and is therefore
deemed to complicated.

Therefore we just return the original wave in this case.