workaround PA_ResetWavesIfRequired did not reset properly on HS removal
The problem is: On HS removal the mode is PLOT_POST_FULL_UPDATE when
entering PA_Update. At the same time all the other settings are constant.
On full update PA_ResetWavesIfRequired is called as well as PA_MarkFailedPulses
but PA_ResetWavesIfRequired does not reset any pulse, since no setting
has changed. So PA_MarkFailedPulses asserts because it wants unmodified
(non-zeroed) pulses.

The workaround is to reset all pulses on full update.