ExtractSweepNumber: Fix and add tests
This is used for extracting the sweep number from a sweep, config, or
its backup waves.

Unfortunately the backup wave parsing failed, resulting in a NaN sweep
number.

The only caller for this case is SWS_DeleteDataWaves. And the fallout in
this case is that a backup wave which should have been deleted, was not
as the comparison NaN >= sweepNo is always false.

Bug present since 0cf35757 (Add helper function for extracting the sweep
number from a sweep/config wave, 2015-07-01).