SWS_DeleteDataWaves: Fix and avoid deleting the wrong sweeps
In 9c27dcd6 (SWS_DeleteDataWaves: Make it faster, 2020-03-26) the author
of this commit introduced some speedups for SWS_DeleteDataWaves.

But these result in the wrong sweep wave being deleted in the common
case that you don't rollback to sweep number zero.

The reason is that the matches wave holds list indizes into `list` with
the wave names to be deleted. This means we need to use matches[i] for
the index into `list` and not i.

The output from the new TestSweepRollback test without this patch looks
like

Entering test case "TestSweepRollback:ITC"
  ITC Device opened, returned deviceID is 0.
  Entering reentry "TestSweepRollback_REENTRY"
  "Config_Sweep_0;Sweep_0;" == "Config_Sweep_2;Sweep_2;" not respecting case: is false.  Assertion "CHECK_EQUAL_STR(refList, list)" failed in line 702 (3931), procedure "UserAnalysisFunctions.ipf"
  Entering reentry "TestSweepRollback_REENTRY_REENTRY"
  "Config_Sweep_0;Config_Sweep_1;Config_Sweep_2;Config_Sweep_3;Sweep_0;Sweep_1;Sweep_2;Sweep_3;" == "Config_Sweep_1;Config_Sweep_2;Config_Sweep_3;Sweep_1;Sweep_2;Sweep_3;" not respecting case: is false.  Assertion "CHECK_EQUAL_STR(refList, list)" failed in line 3966, procedure "UTF_BasicHardwareTests.ipf"
Leaving test case "TestSweepRollback:ITC"
Failed with 2 errors

which shows that instead of all sweeps expect sweep 0, all sweeps except
sweep 2 were deleted.

As we also delete the single sweep folder created by the databrowser
this could also result in the observed case that a config wave is
missing but the sweep wave is present or vice versa.