MSQ_WriteSpikeControlLBNEntries: Correct bugfix for new properties wave
In 73a11a92 (fix for MSQ_WriteSpikeControlLBNEntries to retrieve
diagonality of pulses, 2020-11-17) some changes were made to the way the
properties wave is accessed as the "DiagonalElement" does not exist
anymore.

But that fix was incomplete as it always assumed that failing pulses
were form the current sweep.

This is now fixed.