PA: Speed optimization in PA_LayoutGraphs
In PA_LayoutGraphs the trace plot is set to draw lines between points.
Profiling shows that this uses up a lot of time.
On the other hand drawing lines between points is the Igor default.
So unless a user changes the drawing mode, this change has no effect
but a notable speedup.