DAQ: Add acquisition state handling
In some cases, like adding a labnotebook entry to the current sweep, it
is currently impossible to figure out what the current sweep is, as that
depends if we are during data acquisition, or before or after the sweep.

The new acquisition state variable holds that information.
AS_HandlePossibleAcquisitionStateTransition() takes care of the possible
state transition, calls all necessary code and also checks that the
state transition is faithful during automated testing.