PA: bugfix PA_UpdateScaleBars triggered assert due to operation queue
PA_UpdateScaleBars is called by pushing the command in the operation queue.
For the test code this resulted in the condition that all windows are
closed after a test and then PA_UpdateScaleBars was called.
Since the window the call is referring to was already closed an assertion
was triggered.

Replaced with early abort if referring window does not exist.
Added debug output that triggers for these situations.

Introduced in 426f18f9db1b7136d27db0629a17f42c6236ed2e