PA Tests: Adapted tests for inverted pulse draw order.
In most cases the source of for the profile line extraction was
adjusted.

The wave note check for singleColumnPulseOffset was removed as it does
not exist anymore.

The check for the second sweep in PAT_ImagePLotMultiSweep was tightended,
Interpolate2 is used to simulate the filling by (x) in PA_ShowImage
for the reference data.