PA: With TA the average wave is restricted to common range of all traces
With time alignment enabled the pulses get shifted against each other,
such that their x position and the maximum position matches.
When average was applied to these sets the different count of pulses
at the left and right boundaries resulted in artefacts due to the
different statistics.

Now the average wave is restricted to the range where all pulses have data.
This is automatically carried over to the deconvolution data.