DAP_LockDevice: Throw away ActiveDeviceList for DAQ MD
IP has the quirck that when you duplicate a wave it only duplicates the
dimension labels which are in dimensions with data. And we have one
wave in MIES which can have zero rows, but are still required to have
valid dimension labels.

So let's kill it on locking the first device.

This avoids issues with acquiring new data for the PA_Tests.pxp.