DataBrowser: Added grouped drawing when sweeps are added.
This increases the speed when the DataBrowser graph is updated.

The same logic as in the grouped drawing for the PA plot is used,
with the expension that a structure is introduced to carry the draw
information. This allows to draw the graph elements at a later point in
the code.