GetSizeOfAllWavesInExperiment: Fix sorting by ascending wave sizes
That debug function creates a two column wave, where the second column
holds the size in MB of the wave.

But just using that column for sorting does not work with floating point
values converted to strings. By using the original wavesize wave with
floating point numbers we can fix that.