Overlay Sweeps: Rework headstage removal lists
Up to this commit the approach was to parse the headstage removal list
from the textual listbox wave when that info was needed by
OVS_ParseIgnoreList.

That approach works, but it is slow.

Therefore we now store the active headstages in a separate numeric wave,
GetOverlaySweepHeadstageRemoval, and parse it when it is changed by the
user.