ITC hardware: Avoid passing prepareForDAQ=1 in HW_StopAcq when it is faulty
Since forever we are preparing for the next acquisition when we stop
DAQ, see e.g. e2a4ed93 (Merge pull request #58 in ENG/mies-igor from
bugfix/Incorrect-user-data-during-search to master, 2014-10-07).

And also since forever we have occasionally crashes in the
ITCXOP/ITCXOP2 with the following callstack [1]

itcXOP2-64.xop!AcqCallBack(unsigned int,unsigned int,unsigned __int64,unsigned long,unsigned long)
winmm.dll!DriverCallback()
winmm.dll!timeThread()
kernel32.dll!BaseThreadInitThunk()
ntdll.dll!RtlUserThreadStart()

An investigation now into the ITCMM.DLL machine code using IDA revealed
that the crash happens when dereferencing a 16bit pointer. And the only
16bit wave related to data acquisition is the hardware data wave, see
GetHardwareDataWave().

The crash happens always in AcqCallBack.

This function is called periodically from the Windows OS to transfer
data from the Hardware FIFO to the Software FIFO. This is plausible and
very likley given the documentation in itcmm.pdf in [2], see especially
the DontUseTimerThread setting.

So if we are preparing DAQ again when stopping DAQ, we are
telling the ITCXOP2 that the same config and the *same* data waves can be
used again. But this is actually not true as we might redimension/kill
etc. the config and hardware data waves after stopping data acquisition.

So let's not do that here and hopefully avoid these crashes.

[1]: https://github.com/AllenInstitute/ITCXOP2/issues/13
[2]: https://github.com/AllenInstitute/ITCXOP2