Packages/Testing-MIES/UserAnalysisFunctions.ipf: Mark it as special
We create in MIES a file with the same name for user analysis functions.
This can lead to difficult to diagnose error message when that file is
included instead of the correct one.

Fix that by using a version for the correct file and also have a test
which searched for a marker function.