ScaleToIndexWrapper: Adapt it
The wrapper function wants to iron out all mishandled edge cases of
ScaleToIndex.

With the advent of IP9 some tests now fail. Inspection reveals that IP9
now returns NaN for non-existing dimensions of a wave. But the zeroth
point does exist for any dimensions for a non-empty wave.

So we need to handle this case separately now.