Tests/WB_RegressionTest: Fix test failure in IP9
In IP9 the test case which recreates a pulse train stimset with mixed
frequency fails due to different pulse starting positions.

The error is due to InPlaceRandomShuffle using the default PRNG, which
changed in IP9.

InPlaceRandomShuffle learned the noiseGenMode flag in b28d6a07
(InPlaceRandomShuffle: Add optional parameter determining the RNG type,
2017-09-05) which is also used in WB_PulseTrainSegment, but
unfortunately that commit was buggy as noiseGenMode is not passed to
enoise thus ignoring the passed PRNG.

The fix is twofold. First of all we now don't ignore the noiseGenMode in
InPlaceRandomShuffle anymore, but then also have to pass the IP8 default
PRNG in WB_PulseTrainSegment.