DC_PlaceDataInDAQDataWave: Make the unchanged TP case faster
In TP mode with constant settings we can just reuse the existing
DAQDataWave_TP.

We just need to remember the cache key in the wave note. And if that
matches we just clear the AD data and are done.

No wave versioning is used for the DAQDataWave as this is a purely
optional feature.