DAEphys: Remove decimation factor control
In 30219893 (Add support for decimated oscilloscope wave, 2019-08-29) we
added support for decimated data display. With that addition a
setvariable control was added to tune the decimation factor.

Turns out that this is broken for some values different from the default
values.

Instead of fixing these we now always use the default value of -1.