MSQ_DAScale: Use correct labnotebook prefix for headstage active entry
We want to have MSQ_DA_SCALE and note MSQ_FAST_RHEO_EST. Bug introduced
in e46e3fb7 (Add MSQ_DA_SCALE analysis function, 2018-10-12).

This issue does not have a user impact as the entry is for
internal use relevant only.

And while we are at it, check that entry as well.